Updated directory structure to have client and server directories
